Reimplement tracking of logging client's nick
The original implementation was actually a bit naive, since log open
messages may occur when logs are rotated, not just when the client joins.
This implementation looks for the following sequence of events:

1. Log open message
2. First join message, within 2 seconds of the log open message
3. "Total nicks" message, also within 2 seconds of the log open message

While it is technically possible for this sequence of events to happen
after a log rotate as well (resulting in a false positive), the
likelihood of that - especially within 2 seconds - is very small.

Some irssi clients may choose to filter joins out of logs; this makes
discerning the logging client's initial nick basically impossible,
since its own nick changes don't include the old nick (which is why
this tracking is performed in the first place).  To accommodate
this situation, a `defaultNick` property is added.